Allow floating point scales for LaTeX formulas

+++ Sources/D2Commands/math/LatexRenderer.swift
@@ -16,7 +16,7 @@ class LatexRenderer {
 		cleanUp()
 	}
 	
-	func renderImage(from formula: String, color: String = "white", scale: Int = 6, onError: @escaping (Error) -> Void, then: @escaping (Image) -> Void) throws {
+	func renderImage(from formula: String, color: String = "white", scale: Double = 6, onError: @escaping (Error) -> Void, then: @escaping (Image) -> Void) throws {
 		let timestamp = Int64(Date().timeIntervalSince1970 * 1000000)
 		let outputFile = tempDir.childFile(named: "\(LATEX_PREFIX)-\(timestamp).png")
 
@@ -29,7 +29,7 @@ class LatexRenderer {
 		}
 	}
 	
-	private func renderPNG(from formula: String, to outputFile: TemporaryFile, color: String, scale: Int, onError: @escaping (Error) -> Void, then: @escaping () -> Void) throws {
+	private func renderPNG(from formula: String, to outputFile: TemporaryFile, color: String, scale: Double, onError: @escaping (Error) -> Void, then: @escaping () -> Void) throws {
 		cleanUp()
 		print("Invoking latex-renderer")
 		try node.start(withArgs: [formula, color, outputFile.url.path, String(scale)]) { _ in
